feat：25.2、一级评论总数更新与查询

+++ xiaohashu-comment/xiaohashu-comment-biz/src/main/java/com/quanxiaoha/xiaohashu/comment/biz/domain/dataobject/CommentDO.java
@@ -24,7 +24,7 @@ public class CommentDO {
 
     private String imageUrl;
 
-    private Byte level;
+    private Integer level;
 
     private Long replyTotal;
 
@@ -36,7 +36,7 @@ public class CommentDO {
 
     private Long replyUserId;
 
-    private Byte isTop;
+    private Integer isTop;
 
     private LocalDateTime createTime;
